implement transaction sending with async/await
